getValueの型をuintからfloatに変更

+++ src/application/domain/model/brightness.cpp
@@ -9,15 +9,15 @@ namespace domain::model {
  * @brief 明るさオブジェクトを構築する
  * @param value 明るさの値
  */
-Brightness::Brightness(const uint8_t value) {
+Brightness::Brightness(const float value) {
   mValue = value;
 }
 
 /**
  * @brief 明るさの値を取得します。
- * @return uint8_t 現在の明るさの値
+ * @return float 現在の明るさの値
  */
-[[nodiscard]] uint8_t Brightness::getValue() const noexcept {
+[[nodiscard]] float Brightness::getValue() const noexcept {
   return mValue;
 }
 
@@ -25,7 +25,7 @@ Brightness::Brightness(const uint8_t value) {
  * @brief 明るさを設定する
  * @param value 設定する明るさの値 (0-100)
  */
-void Brightness::setValue(const uint8_t value) {
+void Brightness::setValue(const float value) {
   mValue = value;
 }
